Fix trailing energy gaps, refactor chart options
